Rename `recording-context` to `studio-state`
It contains much more state than only "recording" state
